+++ src/nanodsp/_core.cpp
@@ -16,4 +16,5 @@ NB_MODULE(_core, m) {
     bind_fxdsp(m);
     bind_iirdesign(m);
     bind_paulstretch(m);
+    bind_signalsmith_stretch(m);
 }