+++ src/nanodsp/_core_common.h
@@ -99,3 +99,4 @@ void bind_bloscillators(nb::module_ &m);
 void bind_fxdsp(nb::module_ &m);
 void bind_iirdesign(nb::module_ &m);
 void bind_paulstretch(nb::module_ &m);
+void bind_signalsmith_stretch(nb::module_ &m);